[ui] Reusable Apollo provider
Moves the Apollo provider to a configurable `setupApolloProvider` function.
