+++ ui/src/main.js
@@ -2,16 +2,6 @@ import { createApp } from "vue";
 import App from "./App.vue";
 import router from "./router";
 import { store } from "./store";
-import Cookies from "js-cookie";
-import { createApolloProvider } from "@vue/apollo-option";
-import {
-  ApolloClient,
-  ApolloLink,
-  createHttpLink,
-  InMemoryCache,
-  defaultDataIdFromObject,
-} from "@apollo/client/core";
-import { onError } from "@apollo/client/link/error";
 import { library } from "@fortawesome/fontawesome-svg-core";
 import { FontAwesomeIcon } from "@fortawesome/vue-fontawesome";
 import { far } from "@fortawesome/free-regular-svg-icons";
@@ -20,65 +10,11 @@ import vuetify from "./plugins/vuetify";
 import logger from "./plugins/logger";
 import errorMessages from "./plugins/errors";
 import dateFormatter from "./plugins/dateFormatter";
-
-const AUTHENTICATION_ERROR = "Authentication credentials were not provided";
+import setupApolloProvider from "./apollo/provider";
 
 const API_URL = process.env.VUE_APP_API_URL || `${process.env.BASE_URL}api/`;
 
-// HTTP connection to the API
-const httpLink = createHttpLink({
-  uri: API_URL,
-  credentials: "include",
-});
-
-// Cache implementation
-// Specify object IDs so Apollo can update the cache automatically
-// https://www.apollographql.com/docs/react/v2/caching/cache-configuration/#custom-identifiers
-const cache = new InMemoryCache({
-  dataIdFromObject: (object) => {
-    switch (object.__typename) {
-      case "IndividualType":
-        return object.mk;
-      case "IdentityType":
-        return object.uuid;
-      default:
-        return defaultDataIdFromObject(object);
-    }
-  },
-});
-
-const AuthLink = (operation, next) => {
-  const csrftoken = Cookies.get("csrftoken");
-
-  operation.setContext((context) => ({
-    ...context,
-    headers: {
-      ...context.headers,
-      "X-CSRFToken": csrftoken,
-    },
-  }));
-  return next(operation);
-};
-
-const logoutLink = onError(({ graphQLErrors }) => {
-  if (graphQLErrors && graphQLErrors[0].message == AUTHENTICATION_ERROR) {
-    store.dispatch("logout", {
-      redirect: router.currentRoute?.value?.fullPath,
-    });
-  }
-});
-
-const link = ApolloLink.from([AuthLink, logoutLink, httpLink]);
-
-// Create the apollo client
-const apolloClient = new ApolloClient({
-  link: link,
-  cache,
-});
-
-const apolloProvider = new createApolloProvider({
-  defaultClient: apolloClient,
-});
+const apolloProvider = setupApolloProvider(API_URL, router, store);
 
 createApp(App)
   .component("font-awesome-icon", FontAwesomeIcon)
